applying required changes
